# config/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Customer Management API",
    "version": "1.0.0",
    "description": "This API provides endpoints to manage customer data, including creating, updating, retrieving, and deleting customer information and associated orders.",
    "contact": {
      "name": "API Support",
      "email": "fiorezarn@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/customers/{id}/orders": {
      "post": {
        "tags": ["Customer"],
        "summary": "Create Order",
        "description": "Create an order for a specific customer",
        "operationId": "createOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the customer to create the order for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "completed"
                  },
                  "totalAmount": {
                    "type": "number",
                    "example": 5000
                  }
                }
              },
              "examples": {
                "Create Order": {
                  "value": {
                    "status": "completed",
                    "totalAmount": 5000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order successfully created",
            "content": {
              "application/json": {
                "example": {
                  "message": "success create order",
                  "data": {
                    "id": "order123",
                    "status": "completed",
                    "totalAmount": 5000
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing or invalid data",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid request data"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Customer not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "customer not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Unexpected issue",
            "content": {
              "application/json": {
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
    "/order/{id}": {
      "put": {
        "tags": ["Order"],
        "summary": "Update Order",
        "description": "Update an order by ID",
        "operationId": "updateOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the order to update"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "pending"
                  },
                  "totalAmount": {
                    "type": "number",
                    "example": 5000
                  }
                }
              },
              "examples": {
                "Update Order": {
                  "value": {
                    "status": "pending",
                    "totalAmount": 5000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Order updated successfully."
                    },
                    "order": {
                      "type": "object",
                      "properties": {
                        "orderId": {
                          "type": "integer",
                          "example": 1
                        },
                        "status": {
                          "type": "string",
                          "example": "pending"
                        },
                        "totalAmount": {
                          "type": "number",
                          "example": 5000
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2023-11-12T10:20:30Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Order not found. The specified ID does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Order not found"
                    },
                    "message": {
                      "type": "string",
                      "example": "The order with the specified ID does not exist."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. There was an error updating the order data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "There was an error updating the order data."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Order"],
        "summary": "Delete Order",
        "description": "Delete an order by ID",
        "operationId": "deleteOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the order to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Order deleted successfully."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Order not found. The specified ID does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Order not found"
                    },
                    "message": {
                      "type": "string",
                      "example": "The order with the specified ID does not exist."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. There was an error deleting the order data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "There was an error deleting the order data."
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}